[jsk_recognition_utils] Separate cube.h from geo_util.h
