+++ jsk_recognition_utils/include/jsk_recognition_utils/geo_util.h
@@ -73,6 +73,7 @@
 #include "jsk_recognition_utils/geo/plane.h"
 #include "jsk_recognition_utils/geo/polygon.h"
 #include "jsk_recognition_utils/geo/convex_polygon.h"
+#include "jsk_recognition_utils/geo/cube.h"
 
 // Utitlity macros
 inline void ROS_INFO_EIGEN_VECTOR3(const std::string& prefix,
@@ -248,79 +249,6 @@ namespace jsk_recognition_utils
     ConvexPolygon::Ptr convex_;
     IndexPairSet cells_;
     double resolution_;
-  private:
-    
-  };
-
-  
-  class Cube
-  {
-  public:
-    typedef boost::shared_ptr<Cube> Ptr;
-    Cube(const Eigen::Vector3f& pos, const Eigen::Quaternionf& rot);
-    Cube(const Eigen::Vector3f& pos, const Eigen::Quaternionf& rot,
-         const std::vector<double>& dimensions);
-    Cube(const Eigen::Vector3f& pos, const Eigen::Quaternionf& rot,
-         const Eigen::Vector3f& dimensions);
-    Cube(const Eigen::Vector3f& pos, // centroid
-         const Line& line_a, const Line& line_b, const Line& line_c);
-    virtual ~Cube();
-    std::vector<Segment::Ptr> edges();
-    ConvexPolygon::Ptr intersectConvexPolygon(Plane& plane);
-    std::vector<double> getDimensions() const { return dimensions_; };
-    void setDimensions(const std::vector<double>& new_dimensions) {
-      dimensions_[0] = new_dimensions[0];
-      dimensions_[1] = new_dimensions[1];
-      dimensions_[2] = new_dimensions[2];
-    }
-    jsk_recognition_msgs::BoundingBox toROSMsg();
-
-    /**
-     * @brief
-     * returns vertices as an array of Eigen::Vectro3f.
-     * The order of the vertices is:
-     * [1, 1, 1], [-1, 1, 1], [-1, -1, 1], [1, -1, 1],
-     * [1, 1, -1], [-1, 1, -1], [-1, -1, -1], [1, -1, -1].
-     */
-    Vertices vertices();
-    
-    /**
-     * @brief
-     * returns all the 6 faces as Polygon::Ptr.
-     * TODO: is it should be ConvexPolygon?
-     */
-    std::vector<Polygon::Ptr> faces();
-
-    /**
-     * @brief
-     * compute minimum distance from point p to cube surface.
-     *
-     * Distance computation depends on Polygon::nearestPoint and
-     * this methods just searches a face which resutnrs the smallest
-     * distance.
-     */
-    virtual Eigen::Vector3f nearestPoint(const Eigen::Vector3f& p,
-                                         double& distance);
-  protected:
-    Eigen::Vector3f pos_;
-    Eigen::Quaternionf rot_;
-    std::vector<double> dimensions_;
-
-    /**
-     * @brief
-     * A helper method to build polygon from 4 vertices.
-     */
-    virtual Polygon::Ptr buildFace(const Eigen::Vector3f v0,
-                                   const Eigen::Vector3f v1,
-                                   const Eigen::Vector3f v2,
-                                   const Eigen::Vector3f v3);
-
-    /**
-     * @brief
-     * A helper method to build vertex from x-y-z relatiev coordinates.
-     */
-    virtual Eigen::Vector3f buildVertex(double i, double j, double k);
-    
   private:
     
   };
